Rename util/ to algs/
