+++ src/pegr/gensys/Lua_Interf_Runtime.cpp
@@ -31,7 +31,7 @@
 #include "pegr/gensys/Runtime.hpp"
 #include "pegr/logger/Logger.hpp"
 #include "pegr/script/Script_Util.hpp"
-#include "pegr/util/Algs.hpp"
+#include "pegr/algs/Algs.hpp"
 
 namespace pegr {
 namespace Gensys {
