+++ src/pegr/gensys/Runtime.cpp
@@ -28,7 +28,7 @@
 #include "pegr/logger/Logger.hpp"
 #include "pegr/script/Script_Util.hpp"
 #include "pegr/except/Except.hpp"
-#include "pegr/util/Algs.hpp"
+#include "pegr/algs/Algs.hpp"
 
 namespace pegr {
 namespace Gensys {
